feat: enhance ReposGrid and IssueJumpSection with improved layout, animations, and updated repository stats
